// src/domain/interfaces/metrics-query.interface.ts
export interface MetricsQueryParams {
  dataEmissao?: string;
  representante?: string;
  filial?: string;
  documentoCliente?: string;
  documentoArmazem?: string;
  clienteCodigo?: string;
  codigoProduto?: string;
  descricaoProduto?: string;
  lote?: string;
  notaFiscal?: string;
  paciente?: string;
  dataProcedimento?: string;
  tipoRemessa?: string;
  remessa?: string;
  uf?: string;
  dataVencimento?: string;
}